requirements.txt
matplotlib==3.2.1
seaborn==0.10.1
Pillow==7.2.0
scipy==1.5.2
numpy==1.19.1
pandas==1.1.1
regex==2020.7.14
palettable==3.3.0
requests==2.22.0
scikit_learn==0.23.2
emoji==0.6.0
wordcloud==1.8.0
nltk==3.5
vaderSentiment==3.3.2
